requirements.txt
streamlit==1.31.0
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
langchain==0.1.20
langchain-community==0.0.38
langchain-google-genai==0.0.11
faiss-cpu>=1.8.0
sentence-transformers==2.5.1
google-generativeai>=0.4.1
